Fix sign-conversion warning

+++ src/ex_angle.c
@@ -297,8 +297,8 @@ void extrusion_angle_join (int ncp,		/* number of contour points */
 
    /* malloc the storage we'll need for relaying changed contours to the
     * drawing routines. */
-   mem_anchor =  malloc (2 * 3 * ncp * sizeof(double)
-                      +  2 * 3 * ncp * sizeof(gleDouble));
+   mem_anchor =  malloc (2 * 3 * (size_t) ncp * sizeof(double)
+                      +  2 * 3 * (size_t) ncp * sizeof(gleDouble));
    front_loop = (gleDouble *) mem_anchor;
    back_loop = front_loop + 3 * ncp;
    front_norm = (double *) (back_loop + 3 * ncp);
